requirements.txt
selenium==3.141.0
pytest==6.2.4
python-dotenv==0.18.0
pytest-html==3.1.1